backend/requirements.txt
amadeus==12.0.0
annotated-types==0.7.0
anyio==4.11.0
appnope==0.1.4
APScheduler==3.10.4
asttokens==2.4.1
attrs==25.4.0
blinker==1.7.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
cryptography==46.0.1
debugpy==1.8.1
decorator==5.1.1
dnspython==2.7.0
et_xmlfile==2.0.0
executing==2.0.1
Flask==3.0.2
flask-cors==6.0.1
flatbuffers==25.9.23
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.120.0
google-auth==2.28.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.8.5
google_search_results==2.4.2
googleapis-common-protos==1.63.0
grpcio==1.75.0
grpcio-status==1.62.3
gspread==5.12.4
gspread-dataframe==4.0.0
gunicorn==21.2.0
h11==0.16.0
hypothesis==6.122.3
pytest==8.3.4
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
humanfriendly==10.0
idna==3.10
ImageIO==2.37.2
ipykernel==6.29.4
ipython==8.23.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.1
jupyter_core==5.7.2
lazy_loader==0.4
llvmlite==0.45.1
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numba==0.62.1
numpy>=2.0,<2.3.0
oauthlib==3.3.1
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==24.0
pandas>=2.2.0
parso==0.8.4
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.2.0
pooch==1.8.2
prompt-toolkit==3.0.43
proto-plus==1.26.1
protobuf==4.25.8
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.17.2
PyMatting==1.1.14
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==25.1.2
referencing==0.37.0
rembg==2.0.68
replicate==1.0.7
requests==2.32.5
reportlab>=4.0.0
requests-oauthlib==2.0.0
rpds-py==0.29.0
rsa==4.9.1
ruff==0.13.0
scikit-image==0.25.2
scipy==1.16.3
serpapi==0.1.5
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.10.16
tornado==6.4
tqdm==4.67.1
traitlets==5.14.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.0.1
